requirements.txt
apscheduler
python-dotenv
jsonpath_ng
redis
requests